requirements.txt
aiohttp==3.7.4
aiosignal==1.3.1
akinator==1.1.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.0.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
beautifulsoup4==4.12.2
braceexpand==0.1.7
Brotli==1.1.0
browser-cookie3==0.19.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.0.3
cog-reloader==0.0.2
colorama==0.4.4
contourpy==1.1.0
curl-cffi==0.5.7
cycler==0.11.0
DateTime==5.2
decorator==5.1.1
deep-translator==1.11.4
Deprecated==1.2.14
discord-buttons-plugin==0.1.2
discord-custom-help==0.0.6
discord-protos==0.0.2
discord.py==1.7.3
exceptiongroup==1.1.3
executing==2.0.1
fonttools==4.42.1
frozenlist==1.4.0
g4f==0.3.0.1
google-ai-generativelanguage==0.6.3
google-api-core==2.12.0
google-api-python-client==2.129.0
google-auth==2.23.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
google-generativeai==0.5.3
googleapis-common-protos==1.60.0
googletrans==3.1.0a0
grpcio==1.59.0
grpcio-status==1.59.0
gspread==5.11.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
import-expression==1.1.4
install==1.3.5
ipython==8.24.0
jedi==0.19.1
jishaku==2.3.0
Js2Py==0.74
kiwisolver==1.4.5
langdetect==1.0.9
lz4==4.3.2
matplotlib==3.7.2
matplotlib-inline==0.1.7
multidict==5.2.0
multipledispatch==1.0.0
numpy==1.25.2
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.3
parso==0.8.4
pathlib==1.0.1
pefile==2023.2.7
Pillow==10.0.0
pipdeptree==2.13.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.22.3
protobuf==4.24.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
pydantic==2.7.1
pydantic_core==2.18.2
PyExecJS==1.5.1
Pygments==2.18.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
pyjsparser==2.7.1
pykrx==1.0.45
PyNaCl==1.4.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2023.3
pywin32-ctypes==0.2.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tqdm==4.66.4
traitlets==5.14.3
trio==0.23.1
trio-websocket==0.11.1
tweepy==4.14.0
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==5.0.1
undetected-chromedriver==3.5.4
uritemplate==4.1.1
urllib3==2.0.4
wavelink==0.9.10
wcwidth==0.2.13
websockets==11.0.3
wrapt==1.15.0
wsproto==1.2.0
xlrd==2.0.1
yarl==1.8.2
zope.interface==6.0
